+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - v12: new PCs' tokens are now linked to their actor when created
 - Move links in oracle descriptions now work properly
 - Add Sundered Isles scenes with artwork by AugustBebel on Discord ([#993](https://github.com/ben/foundry-ironsworn/pull/993))
+- Add Sundered Isles structure in preparation for dataforged content. This is in very early stages, and is behind a beta setting. ([#992](https://github.com/ben/foundry-ironsworn/pull/992))
 
 ## 1.23.1